relocate the YoutubePlayerActivity
