+++ ui/tutorial-renderer/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
             android:exported="false"
             android:theme="@style/Theme.GodTools.Activity.Tutorial" />
         <activity
-            android:name=".activity.YoutubePlayerActivity"
+            android:name=".YoutubePlayerActivity"
             android:exported="false"
             android:theme="@style/Theme.GodTools.Modal" />
     </application>
